+++ front/app/containers/Admin/projects/project/permissions/containers/Project/ProjectManagement/ModeratorListRow.tsx
