+++ front/app/containers/Admin/projects/project/permissions/containers/Project/ProjectManagement/index.tsx
@@ -13,8 +13,8 @@ import { Section } from 'components/admin/Section';
 
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
 
-import ModeratorList from '../../../components/ModeratorList';
-import UserSearch from '../../../components/UserSearch';
+import ModeratorList from './ModeratorList';
+import UserSearch from './UserSearch';
 
 import messages from './messages';
